refactor:moved te rate limit utils  to server/common and rename the file
