+++ apps/web/src/api/routes/v1_event.ts
@@ -6,7 +6,7 @@ import isbot from "isbot";
 import { EventService } from "server/services/EventService";
 import { RequestCache } from "server/services/RequestCache";
 import { RequestService } from "server/services/RequestService";
-import { checkRateLimit } from "./utils";
+import { checkRateLimit } from "server/common/ratelimit";
 
 export function makeEventRoute() {
   const app = new Hono().post(
